Extract AudioProcessor and split files
