+++ Sources/LiveKitComponents/UI/Visualizer/AgentBarAudioVisualizer.swift
@@ -14,95 +14,24 @@
  * limitations under the License.
  */
 
-import AVFoundation
+/*
+ * Copyright 2025 LiveKit
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import LiveKit
 import SwiftUI
-
-public class AudioProcessor: ObservableObject, AudioRenderer {
-    public let isCentered: Bool
-    public let smoothingFactor: Float
-
-    // Normalized to 0.0-1.0 range.
-    @Published public var bands: [Float]
-
-    private let _processor: AudioVisualizeProcessor
-    private weak var _track: AudioTrack?
-
-    public init(track: AudioTrack?,
-                bandCount: Int,
-                isCentered: Bool = true,
-                smoothingFactor: Float = 0.3)
-    {
-        self.isCentered = isCentered
-        self.smoothingFactor = smoothingFactor
-        bands = Array(repeating: 0.0, count: bandCount)
-
-        _processor = AudioVisualizeProcessor(bandsCount: bandCount)
-        _track = track
-        _track?.add(audioRenderer: self)
-    }
-
-    deinit {
-        _track?.remove(audioRenderer: self)
-    }
-
-    public func render(pcmBuffer: AVAudioPCMBuffer) {
-        let newBands = _processor.process(pcmBuffer: pcmBuffer)
-        guard var newBands else { return }
-
-        // If centering is enabled, rearrange the normalized bands
-        if isCentered {
-            newBands.sort(by: >)
-            newBands = centerBands(newBands)
-        }
-
-        DispatchQueue.main.async { [weak self] in
-            guard let self else { return }
-
-            self.bands = zip(self.bands, newBands).map { old, new in
-                self._smoothTransition(from: old, to: new, factor: self.smoothingFactor)
-            }
-        }
-    }
-
-    // MARK: - Private
-
-    /// Centers the sorted bands by placing higher values in the middle.
-    @inline(__always) private func centerBands(_ sortedBands: [Float]) -> [Float] {
-        var centeredBands = [Float](repeating: 0, count: sortedBands.count)
-        var leftIndex = sortedBands.count / 2
-        var rightIndex = leftIndex
-
-        for (index, value) in sortedBands.enumerated() {
-            if index % 2 == 0 {
-                // Place value to the right
-                centeredBands[rightIndex] = value
-                rightIndex += 1
-            } else {
-                // Place value to the left
-                leftIndex -= 1
-                centeredBands[leftIndex] = value
-            }
-        }
-
-        return centeredBands
-    }
-
-    /// Applies an easing function to smooth the transition.
-    @inline(__always) private func _smoothTransition(from oldValue: Float, to newValue: Float, factor: Float) -> Float {
-        // Calculate the delta change between the old and new value
-        let delta = newValue - oldValue
-        // Apply an ease-in-out cubic easing curve
-        let easedFactor = _easeInOutCubic(t: factor)
-        // Calculate and return the smoothed value
-        return oldValue + delta * easedFactor
-    }
-
-    /// Easing function: ease-in-out cubic
-    @inline(__always) private func _easeInOutCubic(t: Float) -> Float {
-        t < 0.5 ? 4 * t * t * t : 1 - pow(-2 * t + 2, 3) / 2
-    }
-}
 
 /// A SwiftUI view that visualizes audio levels as a series of vertical bars,
 /// responding to real-time audio data processed from an audio track.
@@ -116,7 +45,8 @@ public class AudioProcessor: ObservableObject, AudioRenderer {
 /// Usage:
 /// ```
 /// let audioTrack: AudioTrack = ...
-/// BarAudioVisualizer(audioTrack: audioTrack)
+/// let agentState: AgentState = ...
+/// AgentBarAudioVisualizer(audioTrack: audioTrack, agentState: agentState)
 /// ```
 ///
 /// - Parameters:
@@ -135,7 +65,7 @@ public class AudioProcessor: ObservableObject, AudioRenderer {
 /// ```
 /// BarAudioVisualizer(audioTrack: audioTrack, barColor: .blue, barCount: 10)
 /// ```
-public struct BarAudioVisualizer: View {
+public struct AgentBarAudioVisualizer: View {
     public let barCount: Int
     public let barColor: Color
     public let barCornerRadius: CGFloat
@@ -153,7 +83,7 @@ public struct BarAudioVisualizer: View {
     @State private var animationTask: Task<Void, Never>?
 
     public init(audioTrack: AudioTrack?,
-                agentState: AgentState? = nil,
+                agentState: AgentState,
                 barColor: Color = .white,
                 barCount: Int = 7,
                 barCornerRadius: CGFloat = 100,
@@ -224,7 +154,7 @@ public struct BarAudioVisualizer: View {
     }
 }
 
-extension BarAudioVisualizer {
+extension AgentBarAudioVisualizer {
     private struct PhaseAnimationProperties {
         typealias HighlightedBars = Set<Int>
